Fixed problem that in AutoResize mode - when the window resizes down, and then back up, the height is not recalculated, and so the video cannot get back to its original size.
Solution - to see if the video container has been given a height already, and if it has, have confidence in the calculated newHeight, and believe that over the parentHeight.

So works as before upon launch, but then as soon as the player has been built - allow the player to resize.